Fix schemaextractor extension in evalutation
